creation of adminController, contactController, ProductsController + index references
