+++ src/Controllers/ContactController.php
@@ -1,0 +1,15 @@
+<?php
+/**
+ * Created by PhpStorm.
+ * User: valerianearon
+ * Date: 28/11/2017
+ * Time: 14:57
+ */
+
+namespace Weasley\Controllers;
+
+
+class ContactController
+{
+
+}
