+++ src/Controllers/DefaultController.php
@@ -42,50 +42,5 @@ class DefaultController extends Controller
         return $this->twig->render('user/contact.html.twig');
     }
 
-    /**
-     * Render login
-     */
-    public function loginAction()
-    {
-        return $this->twig->render('admin/login.html.twig');
-    }
 
-    /**
-     * Render admin
-     */
-    public function adminAction()
-    {
-        return $this->twig->render('admin/admin.html.twig');
-    }
-
-    /**
-     * Render admin contact
-     */
-    public function adminContactAction()
-    {
-        return $this->twig->render('admin/admin_contact.html.twig');/*
-        if ($_SERVER['REQUEST_METHOD'] == "POST") {
-            $errors = [];
-            foreach ($_POST as $key => $value) {
-                if (empty($_POST[$key])) {
-                    $errors[$key] = "Veuillez renseigner le champ " . $key;
-                }
-            }
-
-            if (!empty($errors)) {
-                return $this->twig->render('admin/admin_contact.html.twig', array(
-                    'errors' => $errors
-                ));
-
-            } else {
-                $adresse = $_POST['adresse'];
-                $horaire = $_POST['horaire'];
-                $numero = $_POST['numero'];
-                $commentaire = $_POST['commentaire'];
-                $manager = new ModelManager();
-                $manager->updateContact($adresse, $horaire, $numero, $commentaire);
-            }
-            return $this->twig->render('admin/admin_contact.html.twig');
-        }*/
-    }
 }
